env2dots {nseval}	R Documentation

Copy bindings from an environment into a dots object, or vice versa.

Description

env2dots copies all bindings in the environment (but not those from its parents) into a new dots object. Bindings that are promises will be captured without forcing. Bindings that are not promises will be rendered as forced quotations. The output will not be in any guaranteed order.

Usage

env2dots(
  env = caller(environment()),
  names = ls(envir = env, all.names = TRUE),
  include_missing = TRUE,
  expand_dots = TRUE
)

Arguments

env
An environment.

names
Which names to extract from the environment. By default extracts all bindings present in env, but not in its enclosing environments.

include_missing
Whether to include missing bindings.

expand_dots
Whether to include the contents of ....

Value

A dots object.

[Package nseval version 0.5.1 Index]