format.dots {nseval}	R Documentation

Formatting methods for dots and quotations.

Description

format.dots constructs a string representation of a dots object.

format.quotation constructs a string representation of a quotation object.

format.oneline formats a vector or list so that each item is displayed on one line. It is similar to format.AsIs but tries harder witlanguage objects. The "oneline" class is used by as.data.frame.dots.

Usage

## S3 method for class 'dots'
format(
  x,
  compact = FALSE,
  show.environments = !compact,
  show.expressions = !compact,
  width = NULL,
  ...
)

## S3 method for class 'quotation'
format(
  x,
  compact = FALSE,
  show.environments = !compact,
  show.expressions = !compact,
  width = NULL,
  ...
)

## S3 method for class 'oneline'
format(x, max.width = 50, width = max.width, ...)

## S3 method for class 'dots'
print(x, ...)

## S3 method for class 'quotation'
print(x, ...)

Arguments

x
An object.

compact
Implies show.environments=FALSE and show.expressions=FALSE.

show.environments
Whether to show environments for unforced quotations.

show.expressions
Whether to show expressions for forced quotations.

width
See base::format.

...
Further parameters passed along to base::format.

max.width
See base::format.

[Package nseval version 0.5.1 Index]