get_dots {nseval}	R Documentation

Set or get the contents of ....

Description

get_dots() unpacks ... from a given environment and returns a dots object.

set_dots takes a dots list and uses it to create a binding for ... in a given environment.

Usage

get_dots(env = caller(environment()), inherits = FALSE)

set_dots(env, d, append = FALSE)

Arguments

env
The environment to look in.

inherits
Whether to pull ... from enclosing environments.

d
a [dots] object.

append
if TRUE, the values should be appended to the existing binding. If false, existing binding for "..." will be replaced.

Details

get_dots() is equivalent to dots(...) or arg_list(`...`).

Value

get_dots returns a dots list. If ... is not bound or is missing, it returns an empty dots list.

set_dots returns the updated environment, invisibly.

See Also

env2dots set_arg dots2env

[Package nseval version 0.5.1 Index]